Rename tracking to streamable
The consumer/tracking directory should be renamed to consumer/streamable
 to more closely align the Subscribable/Streamable naming convention

#17-3
