+++ kafka/src/main/java/org/axonframework/extensions/kafka/eventhandling/consumer/streamable/KafkaMessageStream.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.extensions.kafka.eventhandling.consumer.tracking;
+package org.axonframework.extensions.kafka.eventhandling.consumer.streamable;
 
 import org.axonframework.common.Registration;
 import org.axonframework.eventhandling.TrackedEventMessage;
