+++ kafka/src/test/java/org/axonframework/extensions/kafka/eventhandling/consumer/AsyncFetcherTest.java
@@ -23,14 +23,14 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.TopicPartition;
 import org.axonframework.common.AxonConfigurationException;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.Buffer;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.ConsumerUtil;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaEventMessage;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaRecordMetaData;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaTrackingToken;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.SortedKafkaMessageBuffer;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.StreamableKafkaMessageSource;
-import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.TrackingRecordConverter;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.Buffer;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.ConsumerUtil;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.KafkaEventMessage;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.KafkaRecordMetaData;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.KafkaTrackingToken;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.SortedKafkaMessageBuffer;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.StreamableKafkaMessageSource;
+import org.axonframework.extensions.kafka.eventhandling.consumer.streamable.TrackingRecordConverter;
 import org.axonframework.extensions.kafka.eventhandling.producer.ProducerFactory;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.*;
