+++ kafka/src/test/java/org/axonframework/extensions/kafka/eventhandling/consumer/streamable/JSR166TestCase.java
@@ -29,7 +29,7 @@
  * uncompilable code).)
  */
 
-package org.axonframework.extensions.kafka.eventhandling.consumer.tracking;
+package org.axonframework.extensions.kafka.eventhandling.consumer.streamable;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
