+++ kafka/src/test/java/org/axonframework/extensions/kafka/eventhandling/consumer/streamable/SortedKafkaMessageBufferTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.extensions.kafka.eventhandling.consumer.tracking;
+package org.axonframework.extensions.kafka.eventhandling.consumer.streamable;
 
 import java.util.Collection;
 import java.util.LinkedList;
